Consolidate AttributeManager
Changes:
- NaturalAttributeManager constructor takes an initial set of attributes to preload
- RoboticAttributeManager & PersonAttributeManager deleted as they serve no purpose beyong the base class
- Robot & Person updated to use the base Attribute Manager class
- Fixed problem with the bundled crews referencing USA & UK which no longer exist as countries
- Member represents Skills & NaturalAttributes using their enum instead of String
- PersonBuilder remove the need to convert from String to enum
- Move the PersonBuilderImpl to PersonBuilder as the interface provide no advantage
- Add more test for the PersonBuilder

Part of #47
