+++ mars-sim-core/src/main/resources/xml/crew/crew_founders.xml
@@ -8,7 +8,7 @@
 		<person name="Manny Kung" 
 			gender="male" 
 			personality-type="ISTJ"
-			country="USA"
+			country="United States"
 			job="Botanist"
 			favorite-main-dish="Bean Sprout Garlic Stir Fry"
 			favorite-side-dish="Roasted Carrot Soup"
@@ -86,7 +86,7 @@
         <person name="Scott Davis" 
         	gender="male" 
         	personality-type="INFJ"
-			country="USA"
+			country="United States"
         	job="Engineer"
 			favorite-main-dish="Steamed Brown Rice" 
 			favorite-side-dish="Roasted Carrot Soup" 
@@ -138,7 +138,7 @@
        <person name="Barry Evans" 
        		gender="male" 
        		personality-type="INTP"
-			country="UK"
+			country="United Kingdom"
         	job="Doctor"
 			favorite-main-dish="Veggie Sandwich" 
 			favorite-side-dish="Taro Soup"
